- mudar o nome das pasta, todos as pastas devem ter as letras minusculas - no modelo de integração com o repository usar a inserção automática com o  @Resource - implement a interface do service, assim matemos o principio da inversão de dependência, - correção dos métodos chamados de repository
Falta fazer:
- colocar as funções certa para o update, delete(Obs: funções ainda não implementada no service).
- fazer o sistema de consulta personalizada do get
